Fix microtime usage
microtime() returns a string with the components of the time and microtime
making the addition to time() redundant and only functional because of
type coercion. In recent versions of php this treated as a E_NOTICE
causing a significant side effect to the obsevation.

After php 5.4 we can fix this by just passing a parameter to microtime to get
the needed float as a return.

https://www.php.net/microtime#refsect1-function.microtime-parameters